Fix playback.py
The reason why pydub.playback can't play 32bit audios with pyaudio could be that the chunk._data are 32bit integers, but when it's outputted into stream.write(), it needs 32bit floating numbers. "p.get_format_from_width(song.sample_width)" can be seen of getting a called "paFloat32", but when i test it, chunk._data are 32bit integers. So, I fixed the problem by adding a judgment of the sample_width. If it's 4 sample_with, then use pyaudio._portaudio.paInt32. If it's not, then use the origenal method.

This could be the cause of that pydub.AudioSegment.from_file reads audio files by commands of converting them into wav formats, but wav formats are generally not float point numbers' samples. However, the p.get_get_format_from_width() method returns a "paFloat32" when the argument is 4 width. Hard to say whose problem it is, but in short, it's fixed.