Fix playback.py
pydub.playback() can't play 32bit audios with pyaudio correctly because p.get_format_from_width(seg.sample_width) returns "paFloat32" when seg.sample_width is 4, but chunk._data is actually a bunch of 32bit ints, which contradicts the "paFloat32" that means that floats should be given.

So, I fixed the problem by simply adding a judgment of seg.sample_width. If it's not 4, then use the original 
method; If it is 4, then use pyaudio._portaudio.paInt32 to output 32bit ints.